+++ packages/exhibit/src/pavilion/index.ts
@@ -1,1 +1,11 @@
+import { ExhibitContext } from "../common";
+import PavilionContainer from "./pavilionContainer";
+
+ExhibitContext.registerFactory(() => (
+	PavilionContainer.bootstrap()
+));
+
+export { ExhibitContext } from "../common";
+
 export { default as PavilionApp } from "./pavilionApp";
+export { default as PavilionContainer } from "./pavilionContainer";