+++ packages/exhibit/src/pavilion/types.ts
@@ -1,8 +1,16 @@
 import { JSX, LazyExoticComponent } from "react";
 
-import { AppManifest } from "@otuekong-portfolio/features";
+import { CommonModuleEnvironmentKeys } from "@otuekong-portfolio/common";
+import { AppManifest, FeaturesGalleryModuleEnvironmentKeys } from "@otuekong-portfolio/features";
+import { RedisEnvironmentKeys, ResendEnvironmentKeys } from "@otuekong-portfolio/infrastructure-server";
 
 export interface AppDetail {
 	component: LazyExoticComponent<(props?: any) => JSX.Element>;
 	manifest: AppManifest;
 }
+
+export type PavilionModuleEnvironmentKeys = |
+	CommonModuleEnvironmentKeys |
+	FeaturesGalleryModuleEnvironmentKeys |
+	RedisEnvironmentKeys |
+	ResendEnvironmentKeys;